+++ packages/cli/src/loaders/flat.ts
@@ -2,10 +2,7 @@ import { flatten, unflatten } from "flat";
 import { ILoader } from "./_types";
 import { createLoader } from "./_utils";
 
-export default function createFlatLoader(): ILoader<
-  Record<string, any>,
-  Record<string, string>
-> {
+export default function createFlatLoader(): ILoader<Record<string, any>, Record<string, string>> {
   return createLoader({
     pull: async (locale, input) => {
       return flatten(input || {}, {