+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @replexica/react
 
+## 0.3.18
+
+### Patch Changes
+
+- Updated dependencies [[`a2ada16`](https://github.com/lingodotdev/lingo.dev/commit/a2ada16ecf4cd559d3486f0e4756d58808194f7e)]:
+  - @replexica/spec@0.23.0
+
 ## 0.3.17
 
 ### Patch Changes